cleanup and installation instructions

+++ plugins/kanban/custom/updates/migration1019.php
@@ -10,7 +10,7 @@ class Migration1019 extends Migration
         if (Schema::hasTable('users')) {
             Schema::table('users', function($table)
             {
-                $table->integer('team_id')->after('id');
+                $table->integer('team_id')->after('id')->nullable();
                 $table->string('picture')->after('name')->nullable();
             });
         }
@@ -19,11 +19,18 @@ class Migration1019 extends Migration
     public function down()
     {
         if (Schema::hasTable('users')) {
-            Schema::table('users', function($table)
-            {
-                $table->dropColumn('team_id');
-                $table->dropColumn('picture');
-            });
+            if (Schema::hasColumn('users', 'team_id')) {
+                Schema::table('users', function($table)
+                {
+                    $table->dropColumn('team_id');
+                });
+            }
+            if (Schema::hasColumn('users', 'picture')) {
+                Schema::table('users', function($table)
+                {
+                    $table->dropColumn('picture');
+                });
+            }
         }
     }
 }
